Org: Adds optional immediate ticket notifications for all ticket types
TYPE: Feature
LINK: OGC-2124

+++ src/onegov/org/forms/user.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 
+import re
 from functools import cached_property
+from markupsafe import Markup
 from onegov.core.utils import is_valid_yubikey_format
 from onegov.directory.models.directory import Directory
 from onegov.form import Form, merge_forms
@@ -11,10 +13,10 @@ from onegov.form.filters import yubikey_identifier
 from onegov.org import _
 from onegov.org.utils import ticket_directory_groups
 from onegov.ticket import handlers
-from onegov.user import User
+from onegov.user import User, UserGroup
 from onegov.ticket import TicketPermission
 from onegov.user import UserCollection
-from re import match
+from sqlalchemy import and_, or_
 from wtforms.fields import BooleanField
 from wtforms.fields import EmailField
 from wtforms.fields import RadioField
@@ -27,7 +29,6 @@ from wtforms.validators import ValidationError
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from onegov.org.request import OrgRequest
-    from onegov.user import UserGroup
     from wtforms.fields.choices import _Choice
 
 
@@ -37,6 +38,8 @@ AVAILABLE_ROLES = [
     ('supporter', _('Supporter')),
     ('member', _('Member')),
 ]
+
+TICKET_PERMISSION_RE = re.compile(r'(?P<handler>[^-]+)(?:-(?P<group>.+))?')
 
 
 class ManageUserForm(Form):
@@ -199,6 +202,13 @@ class ManageUserGroupForm(Form):
         choices=[],
     )
 
+    immediate_notification = ChosenSelectMultipleField(
+        label=_(
+            'Immediate e-mail notification to members upon ticket submission'
+        ),
+        choices=[],
+    )
+
     directories = ChosenSelectMultipleField(
         label=_('Directories'),
         choices=[],
@@ -224,7 +234,7 @@ class ManageUserGroupForm(Form):
             for u in UserCollection(self.request.session).query()
         ]
         ticket_choices: list[_Choice] = [
-            (f'{key}-', key)
+            (key, key)
             for key in handlers.registry.keys()
         ]
         ticket_choices.extend(
@@ -236,6 +246,116 @@ class ManageUserGroupForm(Form):
             for dir in self.request.session.query(Directory)
         )
         self.ticket_permissions.choices = sorted(ticket_choices)
+
+        # NOTE: We override this in agency with a boolean field
+        if hasattr(self.immediate_notification, 'choices'):
+            self.immediate_notification.choices = sorted(ticket_choices)
+
+    def ensure_exclusive_consistency(self) -> bool:
+        exclusive_permissions = self.exclusive_permissions
+        if not exclusive_permissions:
+            return True
+
+        query = self.request.session.query(
+            TicketPermission.handler_code,
+            TicketPermission.group,
+            UserGroup
+        ).join(UserGroup)
+        if isinstance(self.model, UserGroup):
+            # we can't be inconsistent with ourselves
+            query.filter(TicketPermission.user_group != self.model)
+
+        query.filter(or_(*(
+            and_(
+                TicketPermission.handler_code == handler_code,
+                TicketPermission.group.isnot_distinct_from(group)
+            )
+            for handler_code, group in exclusive_permissions
+        )))
+
+        query = query.filter(TicketPermission.exclusive.is_(False))
+        inconsistencies = query.all()
+        if inconsistencies:
+            assert isinstance(self.ticket_permissions.errors, list)
+            self.ticket_permissions.errors.append(_(
+                'The following selected permissions are invalid '
+                'because immediate notifications are turned on '
+                'in the following groups without a matching '
+                'permission for this ticket:<ul>${inconsistencies}</ul>',
+                mapping={'inconsistencies': Markup('').join(
+                    Markup(
+                        '<li>{permission}: <a href="{url}" '
+                        'target="_blank">{user_group}</a></li>'
+                    ).format(
+                        permission=f'{code}: {group}' if group else code,
+                        url=self.request.link(user_group),
+                        user_group=user_group.title,
+                    )
+                    for code, group, user_group in inconsistencies
+                )},
+                markup=True
+            ))
+            return False
+        return True
+
+    def ensure_immediate_notification_consistency(self) -> bool:
+        immediate_notifications = self.immediate_notifications
+        if not immediate_notifications:
+            return True
+
+        non_exclusive = immediate_notifications - self.exclusive_permissions
+        if not non_exclusive:
+            return True
+
+        query = self.request.session.query(
+            TicketPermission.handler_code,
+            TicketPermission.group
+        )
+        if isinstance(self.model, UserGroup):
+            # we can't be inconsistent with ourselves
+            query.filter(TicketPermission.user_group_id != self.model.id)
+
+        query.filter(or_(*(
+            and_(
+                TicketPermission.handler_code == handler_code,
+                TicketPermission.group.isnot_distinct_from(group)
+            )
+            for handler_code, group in non_exclusive
+        )))
+
+        query = query.filter(TicketPermission.exclusive.is_(True))
+        inconsistencies = query.distinct().all()
+        if inconsistencies:
+            assert isinstance(self.immediate_notification.errors, list)
+            self.immediate_notification.errors.append(_(
+                'The following selected ticket types require '
+                'that ticket permissions are given as well: ${permissions}',
+                mapping={'permissions': ', '.join(
+                    f'{code}: {group}' if group else code
+                    for code, group in inconsistencies
+                )},
+            ))
+            return False
+        return True
+
+    @property
+    def exclusive_permissions(self) -> set[tuple[str, str | None]]:
+        return {
+            (match.group(1), match.group(2))
+            for permission in self.ticket_permissions.data or ()
+            if (match := TICKET_PERMISSION_RE.match(permission))
+        }
+
+    @property
+    def immediate_notifications(self) -> set[tuple[str, str | None]]:
+        if not isinstance(self.immediate_notification.data, list):
+            return set()
+
+        return {
+            (match.group(1), match.group(2))
+            for permission in self.immediate_notification.data
+            if (match := TICKET_PERMISSION_RE.match(permission))
+        }
 
     def update_model(self, model: UserGroup) -> None:
         session = self.request.session
@@ -259,26 +379,43 @@ class ManageUserGroupForm(Form):
         else:
             model.users = []  # type:ignore[assignment]
 
+        exclusive_permissions = self.exclusive_permissions
+        immediate_notifications = self.immediate_notifications
+        missing = exclusive_permissions | immediate_notifications
+
         # Update ticket permissions
         # FIXME: backref across module boundaries
         assert hasattr(model, 'ticket_permissions')
+        permissions = []
+        permission: TicketPermission
         for permission in model.ticket_permissions:
-            session.delete(permission)
-        for permission in self.ticket_permissions.data or ():
-            match_ = match(r'(.[^-]*)-(.*)', permission)
-            if match_:
-                handler_code, group = match_.groups()
-                session.add(
-                    TicketPermission(
-                        handler_code=handler_code,
-                        group=group or None,
-                        user_group=model
-                    )
-                )
-        # initialize meta field with empty dict
-        if not model.meta:
-            model.meta = {}
-        model.meta['directories'] = self.directories.data
+            key = (permission.handler_code, permission.group)
+            exclusive = key in exclusive_permissions
+            notification = key in immediate_notifications
+            if not exclusive and not notification:
+                # no permission object should exist
+                session.delete(permission)
+                continue
+
+            # so we don't add this permission twice
+            missing.discard(key)
+            permission.exclusive = exclusive
+            permission.immediate_notification = notification
+            permissions.append(permission)
+
+        for key in missing:
+            handler_code, group = key
+            permission = TicketPermission(
+                handler_code=handler_code,
+                group=group or None,
+                user_group=model,
+                exclusive=key in exclusive_permissions,
+                immediate_notification=key in immediate_notifications,
+            )
+            session.add(permission)
+            permissions.append(permission)
+
+        model.ticket_permissions = permissions
 
     def apply_model(self, model: UserGroup) -> None:
         self.name.data = model.name
@@ -286,9 +423,14 @@ class ManageUserGroupForm(Form):
         # FIXME: backref across module boundaries
         assert hasattr(model, 'ticket_permissions')
         self.ticket_permissions.data = [
-            f'{permission.handler_code}-{permission.group or ""}'
+            f'{permission.handler_code}-{permission.group}'
+            if permission.group else permission.handler_code
             for permission in model.ticket_permissions
+            if permission.exclusive
         ]
-
-        if model.meta:
-            self.directories.data = model.meta.get('directories', '')
+        self.immediate_notification.data = [
+            f'{permission.handler_code}-{permission.group}'
+            if permission.group else permission.handler_code
+            for permission in model.ticket_permissions
+            if permission.immediate_notification
+        ]
